ContextWindow: PageInspectionPanel #10187
Reorganized the code structure to match the future architecture.
