+++ modules/lib/src/main/resources/assets/js/app/wizard/page/LiveFormPanel.ts
@@ -14,9 +14,8 @@ import type {default as Q} from 'q';
 import {type LiveEditPageInitializationErrorEvent} from '../../../page-editor/event/LiveEditPageInitializationErrorEvent';
 import {type ShowWarningLiveEditEvent} from '../../../page-editor/event/ShowWarningLiveEditEvent';
 import {type LiveEditModel} from '../../../page-editor/LiveEditModel';
-import {cleanupPageEditorBridge, initPageEditorBridge, syncInitialRenderable} from '../../../v6/features/store/pageEditor.store';
-import {clearInspection, inspectItem} from '../../../v6/features/store/pageEditorInspect.store';
-import {cleanupPageInspection, initPageInspectionService} from '../../../v6/features/store/pageInspection.store';
+import {cleanupPageEditorBridge, clearInspection, initPageEditorBridge, inspectItem, syncInitialRenderable} from '../../../v6/features/store/page-editor';
+import {cleanupPageInspection, initPageInspectionService} from '../../../v6/features/store/page-inspection';
 import {$activeWidget} from '../../../v6/features/store/liveViewWidgets.store';
 import {type Content, type ContentBuilder} from '../../content/Content';
 import {type ContentId} from '../../content/ContentId';
