+++ modules/lib/src/main/resources/assets/js/v6/features/store/page-editor/index.ts
@@ -1,0 +1,32 @@
+// Read: computed stores
+export {$hasControllerOrTemplate, $isFragment, $isInsertTabAvailable} from './store';
+export {$isInspecting, $inspectedItem, $inspectedItemType} from './store';
+
+// Read: hooks
+export {usePageState, useContentContext, usePageEditorLifecycle, useDefaultPageTemplateName, useInspectedItem, useInspectedItemType} from './hooks';
+
+// Write: commands
+export {
+    setHasDefaultPageTemplate,
+    syncInitialRenderable,
+    requestSetPageTemplate,
+    requestSetPageController,
+    requestCustomizePage,
+    requestPageReset,
+    executePageReset,
+    requestSetComponentDescriptor,
+    requestUpdateTextComponent,
+    requestSetFragmentComponent,
+    requestComponentAdd,
+    requestComponentRemove,
+    requestComponentDuplicate,
+    requestComponentMove,
+    inspectItem,
+    clearInspection,
+} from './commands';
+
+// Lifecycle
+export {initPageEditorBridge, cleanupPageEditorBridge} from './bridge';
+
+// Types
+export type {PageEditorContentContext, PageEditorLifecycle, InitPageEditorBridgeOptions} from './types';
