+++ modules/lib/src/main/resources/assets/js/v6/features/views/context/widget/page-editor/index.tsx
@@ -1,1 +1,2 @@
-export {PageEditorExtension, PageEditorExtensionElement} from './PageEditorExtension';
+export {PageEditorExtension} from './PageEditorExtension';
+export {PageEditorExtensionElement} from './PageEditorExtensionElement';
