+++ modules/lib/src/main/resources/assets/js/v6/features/views/context/widget/page-editor/inspect/InspectPanel.tsx
@@ -1,7 +1,7 @@
 import {cn} from '@enonic/ui';
 import {useStore} from '@nanostores/preact';
 import type {FC, ReactElement} from 'react';
-import {$inspectedItemType} from '../../../../../store/pageEditorInspect.store';
+import {$inspectedItemType} from '../../../../../store/page-editor';
 import {PageInspectionPanel} from './page';
 import {RegionInspectionPanel} from './region';
