+++ modules/lib/src/main/resources/assets/js/v6/features/views/context/widget/page-editor/inspect/page/index.tsx
@@ -1,2 +1,1 @@
 export {PageInspectionPanel} from './PageInspectionPanel';
-export {PageControllerSelector} from './PageControllerSelector';
