+++ modules/lib/src/main/resources/assets/js/v6/features/views/context/widget/page-editor/inspect/region/RegionInspectionPanel.tsx
@@ -3,7 +3,7 @@ import {Blocks} from 'lucide-react';
 import type {ReactElement} from 'react';
 import {Region} from '../../../../../../../../app/page/region/Region';
 import {ItemLabel} from '../../../../../../shared/ItemLabel';
-import {$inspectedItem} from '../../../../../../store/pageEditorInspect.store';
+import {$inspectedItem} from '../../../../../../store/page-editor';
 import {capitalize} from '../../../../../../utils/format/capitalize';
 
 const REGION_INSPECTION_PANEL_NAME = 'RegionInspectionPanel';
